app/leek/requirements.txt
Flask==1.1.2
flask-restx==0.2.0
boto3==1.9.123
gunicorn==20.0.4
gevent==20.9.0
requests==2.25.0
flask-cors==3.0.7
python-jose[cryptography]==3.1.0
schema==0.7.2
simplejson==3.16.0
elasticsearch==7.8.0
cachecontrol[filecache]==0.12.6

# Additional for agent
kombu==5.0.2
